+++ CHANGELOG.md
@@ -70,7 +70,8 @@
 - Added lazy index constructor and support for `len`, `shape`, `size`, `empty`, `to_pandas()` and `names`. 
 - For `df.index`, Snowpark pandas creates a lazy index object. 
 - For `df.columns`, Snowpark pandas supports a non-lazy version of an Index since the data is already stored locally
-- Added support for Index APIs: `copy()`, `values`, `item()`, `tolist()`, `to_series()` and `to_frame()`
+- Added support for `Index.copy()`
+- Added support for Index APIs: `values`, `item()`, `tolist()`, `to_series()` and `to_frame()`
 
 ## 1.18.0 (2024-05-28)